Editing, adding, etc. works but there is no automatic refresh.  Still quite messy.

+++ writing.h
@@ -1,5 +1,5 @@
 #ifndef WRITING_H
 #define WRITING_H
 int editor_do(char *tmppath, char *defaulttext, char **buf);
-int editor_read(char *path, char **buf);
+long editor_read(char *path, char **buf);
 #endif
